moving cmd line mock
